requirements.txt
absl-py>=1.4.0
gymnasium>=0.28.1
matplotlib>=3.5.3
networkx>=2.8.5
numpy>=1.23.0
pandas==1.5.3
plotly>=5.14.1
pysr>=0.16.3
PyYAML>=6.0
rich>=13.3.1
scikit-learn>=1.1.2
scipy>=1.9.0
stable-baselines3>=2.0
tensorboard>=2.12.0
torch>=2.0.0
torch-geometric>=2.3.0
tqdm>=4.64.0